+++ project/src/components/app/app.tsx
@@ -6,19 +6,8 @@ import NotFoundPage from '../../pages/not-found-page/not-found-page';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import PrivateRoute from '../private-route/private-route';
 import { AppRoute } from '../../const';
-import { useEffect } from 'react';
-import { useAppDispatch } from '../../hooks';
-import { checkAuthAction, fetchFavoriteRooms, fetchRooms } from '../../store/api-actions';
 
 function App(): JSX.Element {
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(fetchFavoriteRooms());
-    dispatch(fetchRooms());
-    dispatch(checkAuthAction());
-  }, [dispatch]);
-
   return (
     <BrowserRouter>
       <Routes>
