+++ project/src/components/user-information/user-information.tsx
@@ -1,0 +1,43 @@
+import { Fragment, useEffect } from 'react';
+import { Link } from 'react-router-dom';
+import { AppRoute } from '../../const';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+import { store } from '../../store';
+import { fetchFavoriteRooms, logoutAction } from '../../store/api-actions';
+import { getFavoriteRooms } from '../../store/favorites/selectors';
+import { getLoggedUser } from '../../store/user/selectors';
+
+export function UserInformation(): JSX.Element {
+  const loggedUser = useAppSelector(getLoggedUser);
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    dispatch(fetchFavoriteRooms());
+  }, [dispatch]);
+
+  const favoriteRooms = useAppSelector(getFavoriteRooms);
+
+
+  return (
+    <Fragment>
+      <li className="header__nav-item user">
+        <Link className={'header__nav-link header__nav-link--profile'} to={AppRoute.Favorites}>
+          <div className="header__avatar-wrapper user__avatar-wrapper">
+            <img className="header__avatar user__avatar" src={loggedUser?.avatarUrl} width="74" height="74" alt="Host avatar"/>
+          </div>
+          <span className="header__user-name user__name">{loggedUser?.email}</span>
+          <span className="header__favorite-count">{favoriteRooms.length}</span>
+        </Link>
+      </li>
+      <li className="header__nav-item">
+        <a className="header__nav-link" href="/#" onClick={(evt)=> {
+          evt.preventDefault();
+          store.dispatch(logoutAction());
+        }}
+        >
+          <span className="header__signout">Sign out</span>
+        </a>
+      </li>
+    </Fragment>
+  );
+}
