+++ project/src/index.tsx
@@ -5,6 +5,10 @@ import { Provider } from 'react-redux';
 import { store } from './store';
 import {ToastContainer} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import { checkAuthAction, fetchRooms } from './store/api-actions';
+
+store.dispatch(fetchRooms());
+store.dispatch(checkAuthAction());
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
